Add tests for the fulfillment details components (#13087)
* initial commit

* secondary commit

* write some tests

* add some more fulfillment details form tests

* update naming after rebase

* fix addressForm helper

* fix type errors

* use new and legacy address form helpers

---------
