+++ src/Apps/Order/Routes/Shipping2/__tests__/Shipping2.jest.tsx
@@ -30,7 +30,7 @@ import { useFeatureFlag } from "System/useFeatureFlag"
 import {
   fillAddressForm,
   validAddress,
-} from "Components/__tests__/Utils/addressForm"
+} from "Components/__tests__/Utils/addressForm2"
 import userEvent from "@testing-library/user-event"
 import { flushPromiseQueue } from "DevTools/flushPromiseQueue"
 import { queryByAttribute } from "@testing-library/dom"
@@ -70,7 +70,7 @@ jest.mock("System/useFeatureFlag", () => ({
 
 const mockShowErrorDialog = jest.fn()
 jest.mock("Apps/Order/Dialogs", () => ({
-  ...jest.requireActual("../../Dialogs"),
+  ...jest.requireActual("Apps/Order/Dialogs"),
   injectDialog: Component => props => (
     <Component {...props} dialog={{ showErrorDialog: mockShowErrorDialog }} />
   ),
@@ -152,7 +152,7 @@ const meWithAddresses: Shipping2TestQuery$rawResponse["me"] = Object.assign(
 )
 
 const saveAndContinue = async () => {
-  await userEvent.click(screen.getByText("Save and Continue"))
+  await waitFor(() => userEvent.click(screen.getByText("Save and Continue")))
 }
 
 const recommendedAddress = {
