+++ src/Components/__tests__/Utils/addressForm2.ts
@@ -1,4 +1,8 @@
-/* Address utilities for the old address form without improvements like state select */
+/**
+ * Address utilities for the new address form (Used in Shipping2)
+ * with improvements like state select
+ */
+
 import { Address } from "Components/Address/AddressForm"
 import { CountrySelect } from "Components/CountrySelect"
 import { Input } from "@artsy/palette"
@@ -48,23 +52,32 @@ export const fillAddressForm = async (address: Address) => {
     const line1Input = screen.getByPlaceholderText("Street address")
     expect(line1Input).toBeEnabled()
   })
-  const name = screen.getByPlaceholderText("Full name")
   const country = screen.getByTestId("AddressForm_country")
+  await userEvent.selectOptions(country, [address.country])
+
+  const name = screen.getByPlaceholderText("Full name")
   const addressLine1 = screen.getByPlaceholderText("Street address")
   const addressLine2 = screen.getByPlaceholderText("Apt, floor, suite, etc.")
   const city = screen.getByPlaceholderText("City")
-  const region = screen.getByPlaceholderText("State, province, or region")
-  const postalCode = screen.getByPlaceholderText("ZIP/postal code")
+  const region = screen.getByPlaceholderText(
+    address.country === "US" ? "State" : "State, province, or region"
+  )
+  const postalCode = screen.getByPlaceholderText(
+    address.country === "US" ? "ZIP code" : /ZIP\/postal code/,
+    { exact: false }
+  )
   const phoneNumber = screen.getAllByPlaceholderText(
     "Add phone number including country code"
   )[0]
 
-  userEvent.paste(name, address.name)
-  userEvent.selectOptions(country, [address.country])
-  userEvent.paste(addressLine1, address.addressLine1)
-  userEvent.paste(addressLine2, address.addressLine2)
-  userEvent.paste(city, address.city)
-  userEvent.paste(region, address.region)
-  userEvent.paste(postalCode, address.postalCode)
-  userEvent.paste(phoneNumber, address.phoneNumber!)
+  await Promise.all([
+    userEvent.paste(name, address.name),
+    userEvent.paste(addressLine1, address.addressLine1),
+    userEvent.paste(addressLine2, address.addressLine2),
+    userEvent.paste(city, address.city),
+    userEvent.paste(region, address.region),
+    userEvent.paste(postalCode, address.postalCode),
+  ])
+  address.phoneNumber &&
+    (await userEvent.paste(phoneNumber, address.phoneNumber))
 }
